requirements.txt
streamlit
google-generativeai
python-dotenv
requests
fastapi
uvicorn
mysql-connector-python
bcrypt